+++ frontend/game/ui-steps/message-step.tsx
@@ -1,20 +1,12 @@
-import {bind} from 'bind-decorator';
 import * as React from 'react';
 
-import { AnyUiState, StepType, MessageStepUiState } from '../../../common/game';
+import { MessageStepUiState } from '../../../common/game';
 
 
 interface Props extends MessageStepUiState {
 }
-interface State {
-}
 
-export class MessageStep extends React.PureComponent<Props, State> {
-    constructor(props: Props) {
-        super(props);
-        this.state = {};
-    }
-
+export class MessageStep extends React.PureComponent<Props> {
     public render() {
         return (
             <div className={"chat-segment" + (this.props.character ? ` ${this.props.character}` : '')}>
